// Client/src/app/core/modules/openapi/model/reviewer-pre-submission-guideline-write-dto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ReviewerPreSubmissionGuidelineWriteDto { 
    reviewerRole: ReviewerPreSubmissionGuidelineWriteDto.ReviewerRoleEnum;
    title: string;
    content: string;
    goodExample?: string;
    badExample?: string;
    relatedModuleFieldKey?: string;
    sortOrder?: number;
    active?: boolean;
}
export namespace ReviewerPreSubmissionGuidelineWriteDto {
    export type ReviewerRoleEnum = 'ADMIN' | 'QUALITY_MANAGEMENT' | 'ACADEMIC_PROGRAM_ADVISOR' | 'EXAMINATION_BOARD' | 'PROFESSOR' | 'PROGRAM_COORDINATOR' | 'SPECIALIZATION_AREA_COORDINATOR';
    export const ReviewerRoleEnum = {
        Admin: 'ADMIN' as ReviewerRoleEnum,
        QualityManagement: 'QUALITY_MANAGEMENT' as ReviewerRoleEnum,
        AcademicProgramAdvisor: 'ACADEMIC_PROGRAM_ADVISOR' as ReviewerRoleEnum,
        ExaminationBoard: 'EXAMINATION_BOARD' as ReviewerRoleEnum,
        Professor: 'PROFESSOR' as ReviewerRoleEnum,
        ProgramCoordinator: 'PROGRAM_COORDINATOR' as ReviewerRoleEnum,
        SpecializationAreaCoordinator: 'SPECIALIZATION_AREA_COORDINATOR' as ReviewerRoleEnum
    };
}